Consolidate and rename top-level modules
This consolidates and clarifies top level module names.

functions/ only dealt with builtin functions so it's been renamed.
interfaces/ only dealt with builtin interfaces so it's been renamed.

codegen/ was going to be a module to replace parser, but we are going to
replace parser/ entirely with another pass (name tbd), so throw it all
in together.
parser/ does mostly codegen at this point but because it is all going to
be replaced with new(TM) code, rename it old_codegen to mark it as for
deletion.
types/ is the old way of dealing with types so throw it into old_codegen

signatures/ had too much ast code so it went in ast/.

settings.py seemed like it belonged in compiler/
opcodes seemed like it belonged in a new folder for evm specifics.
compile_lll, s_expressions and optimizer seemed to belong in a new lll/
folder.

context/ was intended as the next pass after parsing, so rename it to
semantics/ to reflect that it's a pass dedicated to analysing the
semantics of user code.
semantics/types/meta/ was renamed semantics/types/user/ to reflect that
they are user-defined types